requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
langchain
langchain-community
langchain-ollama
pypdf
celery
redis
python-dotenv
python-multipart
alembic
pytest
pytest-asyncio
httpx
pytest-cov